tests: Reorganize structure of the integration tests
Create a testlib sub-package for test helpers and an example sub-package
for example tests.
Place the current no-mac-spoofing example in its own well named test
file.

Future examples are expected to be placed under the `example`
sub-package with a dedicated test file name.
